+++ src/powerpwn/models/flow_arguments.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel, Field
 from enum import Enum
-from typing import Optional
 
 
 class FlowToRunEnum(Enum):
